use separate files for BridgeSSLClient
