+++ src/BridgeSSLClient.h
@@ -1,0 +1,36 @@
+/*
+  Copyright (c) 2013 Arduino LLC. All right reserved.
+
+  This library is free software; you can redistribute it and/or
+  modify it under the terms of the GNU Lesser General Public
+  License as published by the Free Software Foundation; either
+  version 2.1 of the License, or (at your option) any later version.
+
+  This library is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+  Lesser General Public License for more details.
+
+  You should have received a copy of the GNU Lesser General Public
+  License along with this library; if not, write to the Free Software
+  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
+*/
+
+#ifndef _BRIDGE_SSL_CLIENT_H_
+#define _BRIDGE_SSL_CLIENT_H_
+
+#include <Bridge.h>
+#include <Client.h>
+#include <BridgeClient.h>
+
+class BridgeSSLClient : public BridgeClient {
+  public:
+    // Constructor with a user provided BridgeClass instance
+    BridgeSSLClient(uint8_t _h, BridgeClass &_b = Bridge);
+    BridgeSSLClient(BridgeClass &_b = Bridge);
+    ~BridgeSSLClient();
+
+    virtual int connect(const char* host, uint16_t port);
+};
+
+#endif // _BRIDGE_SSL_CLIENT_H_
